Runs: standalone /runs/detail/ fullscreen view + panel-header polish
Adds a real fullscreen detail route behind the "Full view ↗" button so
clicking it opens a focused single-run view instead of the same
split-pane page with a `?run=<id>` anchor. Fixes the text-wrap bug that
made the two panel-header buttons collapse into ragged two-line chips
when the title was long.

Panel header fixes

- Title column now gets `min-w-0 flex-1` + 2-line `line-clamp` so long
  titles (file paths, long GitHub issue titles) stop squeezing the
  action buttons.
- Buttons get `shrink-0` + `whitespace-nowrap` so they stay on a clean
  single line no matter how narrow the panel is.
- Button labels tightened to "⌕ Receipt" and "Full view ↗" with
  descriptive `title` attrs for tooltip hover; the old "⌕ Receipt
  preview" / "Open in new tab ↗" strings blew out the header at
  split-pane widths.

Shared fixtures

- New `components/runs/fixtures.ts` centralises FIXTURE_RUN_ROWS,
  FIXTURE_RECOMMENDATIONS, FIXTURE_FILTERS, FIXTURE_JOB_DEFINITIONS,
  FIXTURE_LIFECYCLE — used by both the queue page (fallback data when
  the live API is empty) and the detail page.

Self-contained LoadedRunView

- New `components/runs/LoadedRunView.tsx` owns everything a single-run
  detail view needs: live/fixture row lookup, job-definition fetch,
  GitHub context build, submit handler, receipt-drawer state, and the
  big chunk of fixture verifier / settle / lifecycle copy that used to
  live inline in runs/page.tsx. Takes a `runId` prop and optionally
  hides the lifecycle rail when the host already renders one.
- runs/page.tsx is now ~100 lines (was ~700) — just the queue + filter
  + recommendation logic, with LoadedRunView in the sticky right
  column. Lifecycle rail + recommendation strip stay below the split
  pane as before.

Standalone detail route

- New `/runs/detail/?id=<runId>` page renders the same LoadedRunView
  at full width (max-w-1100) with a "← Back to queue" header and a
  "run <id> · fullscreen view" label on the right.
- Uses `useSearchParams` inside a Suspense boundary so it plays nice
  with Next's static export on the VPS. Query param instead of a
  dynamic `[id]` segment keeps the static export straightforward.
- At this width the panel's `@container` query flips its inner grid to
  2-col — stake + job context on the left, verifier + settlement on
  the right. Use-case: focused dispute review, sharing a direct link
  to a cosigner, multi-tab triage.
- "Full view ↗" in the main page's panel header now links to
  `/runs/detail/?id=<selectedId>` (target=_blank). On the detail page
  itself, `standaloneUrl` is omitted so the button hides — no "go to
  the page you're already on" bug.
